+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # v2.3.0
 
+* added remember-author feature across logouts (requires js)
 * added default_author config.ini variable
 * added integration tests for all supported PHP versions
 * fixed test:unit / missing docs in release zip/tgz